requirements.txt
ipython
click
grpcio-tools
flake8
